requirements.txt
utils-nuuuwan
Pillow
scikit-learn